doc: generate doxygen html output from the scanner
This switches the scanner to generate doxygen-compatible tags for the
generated protocol headers, and hooks up the doxygen build to generate server
and client-side API documentation.

For the wayland protocol, this generates a mainpage with the copyright
information, all interfaces are separated by doxygen groups and thus listed in
"Modules" in the generated output.

Function, struct and #defines are documented in doxygen style and associated
with the matching interface.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>